fix broadcast size in test

+++ tests/api/test_tt_device.cpp
@@ -267,7 +267,7 @@ TEST(ApiTTDeviceTest, BroadcastIO) {
     std::vector<int> pci_device_ids = PCIDevice::enumerate_devices();
 
     uint64_t address = 0x0;
-    std::vector<uint8_t> data_write = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+    std::vector<uint32_t> data_write = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
 
     for (int pci_device_id : pci_device_ids) {
         std::unique_ptr<TTDevice> tt_device = TTDevice::create(pci_device_id);
@@ -279,14 +279,14 @@ TEST(ApiTTDeviceTest, BroadcastIO) {
         const std::vector<CoreCoord>& tensix_cores = soc_desc.get_cores(CoreType::TENSIX, CoordSystem::TRANSLATED);
 
         // Zero out all tensix cores before broadcasting.
-        std::vector<uint8_t> zeros(data_write.size(), 0);
+        std::vector<uint32_t> zeros(data_write.size(), 0);
         for (const CoreCoord& core : tensix_cores) {
             tt_device->write_to_device(zeros.data(), core, address, zeros.size());
         }
 
         // Verify zeros landed.
         for (const CoreCoord& core : tensix_cores) {
-            std::vector<uint8_t> readback(data_write.size(), 1);
+            std::vector<uint32_t> readback(data_write.size(), 1);
             tt_device->read_from_device(readback.data(), core, address, readback.size());
             ASSERT_EQ(zeros, readback) << "Core " << core.str() << " on chip " << pci_device_id
                                        << " should have been zeroed before the broadcast write.";
@@ -296,7 +296,7 @@ TEST(ApiTTDeviceTest, BroadcastIO) {
 
         // All tensix cores should now have the broadcast data.
         for (const CoreCoord& core : tensix_cores) {
-            std::vector<uint8_t> readback(data_write.size());
+            std::vector<uint32_t> readback(data_write.size());
             tt_device->read_from_device(readback.data(), core, address, readback.size());
             ASSERT_EQ(data_write, readback) << "Core " << core.str() << " on chip " << pci_device_id
                                             << " should have received the broadcast write.";
